manage_shop: add category revisions delete

+++ admin_modules/yf_category_editor.class.php
@@ -43,7 +43,6 @@ class yf_category_editor {
 				'cats__prepare_for_box__'.$cat_info['name'].'_1_0',
 			));
 		}
-		_class( '_shop_categories', 'modules/shop/' )->_refresh_cache();
 	}
 
 	/**
@@ -658,15 +657,21 @@ class yf_category_editor {
 	/**
 	*/
 	function delete_item() {
-		$_GET['id'] = intval($_GET['id']);
-		if (!empty($_GET['id'])) {
-			$item_info = db()->query_fetch('SELECT * FROM '.db('category_items').' WHERE id='.intval($_GET['id']));
-		}
-		if (!empty($item_info)) {
-			$cat_info = db()->query_fetch('SELECT * FROM '.db('categories').' WHERE id='.intval($item_info['cat_id']));
-			db()->query('DELETE FROM '.db('category_items').' WHERE id='.intval($_GET['id']));
-			common()->admin_wall_add(array('category item deleted: '.$item_info['id'], $_GET['id']));
-			module('category_editor')->_purge_category_caches($cat_info);
+		$id = (int)$_GET['id']; $_GET['id'] = $id;
+		$object = $_GET[ 'object' ];
+		$action = $_GET[ 'action' ];
+		if( $id < 1 ) { return js_redirect( './?object='.$object, 'item id < 1' ); }
+		$db_item   = db( 'category_items' );
+		$item_info = db()->query_fetch( 'SELECT * FROM ' . $db_item . ' WHERE id = ' . $id );
+		if( !empty( $item_info ) ) {
+			$db       = db( 'categories' );
+			$cats_id = $item_info[ 'cat_id' ];
+			$cat_info = db()->query_fetch( 'SELECT * FROM ' . $db . ' WHERE id = ' . (int)$cats_id );
+			_class( 'core_events' )->fire( 'category_editor.delete_item.before', array( $id, $cats_id ) );
+				db()->query('DELETE FROM '.db('category_items').' WHERE id='.intval($_GET['id']));
+				common()->admin_wall_add(array('category item deleted: '.$item_info['id'], $_GET['id']));
+				$this->_purge_category_caches( $cat_info );
+			_class( 'core_events' )->fire( 'category_editor.delete_item.after', array( $id, $cats_id ) );
 		}
 		if ($_POST['ajax_mode']) {
 			main()->NO_GRAPHICS = true;
